feat(backend)!: Adds Authorization Handling to all Request Handlers
Now Authorization gets checked in every request handler; For Querying multiple resources a authorization filter is created; in other cases exceptions are thrown when unauthorized

+++ src/backend/ProjectMetadataPlatform.Application/OfficeLocations/DeleteOfficeLocationCommandHandler.cs
@@ -2,6 +2,8 @@
 using System.Threading.Tasks;
 using MediatR;
 using ProjectMetadataPlatform.Application.Interfaces;
+using ProjectMetadataPlatform.Domain.Authorization;
+using ProjectMetadataPlatform.Domain.Errors.AuthorizationExceptions;
 using ProjectMetadataPlatform.Domain.Logs;
 
 namespace ProjectMetadataPlatform.Application.OfficeLocations;
@@ -14,6 +16,7 @@ public class DeleteOfficeLocationCommandHandler : IRequestHandler<DeleteOfficeLo
     private readonly IOfficeLocationRepository _officeLocationRepository;
     private readonly ILogRepository _logRepository;
     private readonly IUnitOfWork _unitOfWork;
+    private readonly IAuthorizationService _authorizationService;
 
     /// <summary>
     /// Creates a new instance of <see cref="DeleteOfficeLocationCommandHandler" />.
@@ -21,12 +24,14 @@ public class DeleteOfficeLocationCommandHandler : IRequestHandler<DeleteOfficeLo
     public DeleteOfficeLocationCommandHandler(
         IOfficeLocationRepository officeLocationRepository,
         ILogRepository logRepository,
-        IUnitOfWork unitOfWork
+        IUnitOfWork unitOfWork,
+        IAuthorizationService authorizationService
     )
     {
         _officeLocationRepository = officeLocationRepository;
         _logRepository = logRepository;
         _unitOfWork = unitOfWork;
+        _authorizationService = authorizationService;
     }
 
     /// <summary>
@@ -41,6 +46,17 @@ public class DeleteOfficeLocationCommandHandler : IRequestHandler<DeleteOfficeLo
     )
     {
         var location = await _officeLocationRepository.GetOfficeLocationAsync(request.Id);
+        if (
+            !(
+                await _authorizationService.CheckAccess(
+                    location,
+                    [AuthorizationConstants.Actions.DELETE]
+                )
+            )[AuthorizationConstants.Actions.DELETE]
+        )
+        {
+            throw new UnauthorizedException();
+        }
         _ = await _officeLocationRepository.DeleteOfficeLocationAsync(location);
         await _logRepository.AddOfficeLocationLogForCurrentActor(
             location,
