+++ src/backend/ProjectMetadataPlatform.Application/Plugins/DeleteGlobalPluginCommandHandler.cs
@@ -3,6 +3,8 @@ using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
 using ProjectMetadataPlatform.Application.Interfaces;
+using ProjectMetadataPlatform.Domain.Authorization;
+using ProjectMetadataPlatform.Domain.Errors.AuthorizationExceptions;
 using ProjectMetadataPlatform.Domain.Errors.PluginExceptions;
 using ProjectMetadataPlatform.Domain.Logs;
 
@@ -16,6 +18,7 @@ public class DeleteGlobalPluginCommandHandler : IRequestHandler<DeleteGlobalPlug
     private readonly IPluginRepository _pluginRepository;
     private readonly ILogRepository _logRepository;
     private readonly IUnitOfWork _unitOfWork;
+    private readonly IAuthorizationService _authorizationService;
 
     /// <summary>
     /// Creates a new instance of<see cref="DeleteGlobalPluginCommandHandler"/>.
@@ -23,15 +26,18 @@ public class DeleteGlobalPluginCommandHandler : IRequestHandler<DeleteGlobalPlug
     /// <param name="pluginRepository"></param>
     /// <param name="logRepository"></param>
     /// <param name="unitOfWork"></param>
+    /// <param name="authorizationService"></param>
     public DeleteGlobalPluginCommandHandler(
         IPluginRepository pluginRepository,
         ILogRepository logRepository,
-        IUnitOfWork unitOfWork
+        IUnitOfWork unitOfWork,
+        IAuthorizationService authorizationService
     )
     {
         _pluginRepository = pluginRepository;
         _logRepository = logRepository;
         _unitOfWork = unitOfWork;
+        _authorizationService = authorizationService;
     }
 
     /// <summary>
@@ -48,6 +54,18 @@ public class DeleteGlobalPluginCommandHandler : IRequestHandler<DeleteGlobalPlug
         var plugin =
             await _pluginRepository.GetPluginByIdAsync(request.Id)
             ?? throw new PluginNotFoundException(request.Id);
+
+        if (
+            !(
+                await _authorizationService.CheckAccess(
+                    plugin,
+                    [AuthorizationConstants.Actions.DELETE]
+                )
+            )[AuthorizationConstants.Actions.DELETE]
+        )
+        {
+            throw new UnauthorizedException();
+        }
         if (plugin.IsArchived)
         {
             _ = await _pluginRepository.DeleteGlobalPlugin(plugin);
